search-app/requirements.txt
# ADK Framework
google-adk>=0.1.0

# A2A Protocol SDK
a2a-sdk>=0.2.0

# Additional dependencies for streaming and async operations
aiohttp>=3.8.0
asyncio
python-dotenv>=0.19.0

# Optional: For enhanced functionality
structlog>=21.0.0
prometheus-client>=0.12.0 